+++ src/lib/api/index.ts
@@ -4,3 +4,4 @@ export * as inviteCodes from "./inviteCodes";
 export * as spaces from "./spaces";
 export * as resources from "./resources";
 export * as requests from "./requests";
+export * as access from "./access";
